+++ CHANGELOG.md
@@ -277,7 +277,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Metavariables were not matching due to go parser adding empty statements in golang
 
 
-## [0.4.8](https://github.com/returntocorp/sgrep/releases/tag/0.4.8) - 2020-03-09
+## [0.4.8](https://github.com/returntocorp/semgrep/releases/tag/0.4.8) - 2020-03-09
 
 ### Added
 * Constant propagation for some langauges. Golang example: